+++ src/preProcessing/Homography/homography.c
@@ -1,6 +1,6 @@
 #include <SDL2/SDL.h>
 #include "SDL2/SDL_image.h"
-#include "../SDL_Function/sdlFunction.h"
+#include "preProcessing/SDL_Function/sdlFunction.h"
 
 #include <err.h>
 
@@ -160,8 +160,8 @@ SDL_Surface* Homography_Transform(SDL_Surface* image, int size)
 				Uint32* new_pixels = (Uint32*)new_img->pixels;
 				Uint32* pixels = (Uint32*)image->pixels;
 
-				Uint32 pixel = pixels[y * image->w + x];
-				new_pixels[j + new_img->w + i] = pixel;
+				Uint32 pixel = pixels[(int)y * image->w + (int)x];
+				new_pixels[(int)j + new_img->w + (int)i] = pixel;
 			}
 		}
 	}
